refactoring, and changing localisation files, etc
