+++ src/pages/PrivatePolicyHome/PrivatePolicyHome.tsx
@@ -1,0 +1,12 @@
+import React, { FunctionComponent } from 'react';
+import { PrivacyPolicy } from './PrivatePolicy/PrivatePolicy';
+import { Header } from '../Home/Header/Header';
+import { Footer } from '../Home/Footer/Footer';
+
+export const PrivacyPolicyHome: FunctionComponent = () => (
+  <>
+    <Header />
+    <PrivacyPolicy />
+    <Footer />
+  </>
+);
